modularize src
