+++ src/main.rs
@@ -1,30 +1,35 @@
 use colored::*;
 use log::{debug, error, info}; // trace, warn
-mod logger;
 use std::io::{self, Write};
+
+mod api;
+mod cli;
+mod context;
+mod helper;
+mod logger;
 
 #[tokio::main]
 async fn main() {
-    let log_path = gtc::set_home_dir_path(".gtc.log");
+    let log_path = helper::set_home_dir_path(".gtc.log");
     let log_config = logger::setup_logger(log_path);
     let _handle = log4rs::init_config(log_config).unwrap();
 
     const PROFILE: &str = ".gtc";
-    let profile_path = gtc::set_home_dir_path(PROFILE);
+    let profile_path = helper::set_home_dir_path(PROFILE);
     debug!("profile_path - {}", profile_path.to_string_lossy());
 
-    let matches = gtc::cli().get_matches();
+    let matches = cli::cli().get_matches();
 
     if matches.contains_id("message") {
         let mut ctx = if profile_path.exists() {
-            let mut ctx_read = gtc::read_context(&profile_path);
+            let mut ctx_read = context::read_context(&profile_path);
 
             match ctx_read.openai_key.is_empty() {
                 // prompt user for openai key
                 true => {
                     info!("No OpenAI API key found");
                     // get openai key from user
-                    let openai_key = gtc::input(
+                    let openai_key = helper::input(
                         "No OpenAI API key found, please enter:",
                         &mut io::stdin().lock(),
                         &mut io::stdout(),
@@ -38,19 +43,19 @@ async fn main() {
             }
         } else {
             // create profile if it doesn't exist and prompt user for openai key
-            let openai_key = gtc::input(
+            let openai_key = helper::input(
                 "No OpenAI API key found, please enter:",
                 &mut io::stdin().lock(),
                 &mut io::stdout(),
             );
             // update context and return
-            gtc::Context {
+            context::Context {
                 openai_key: openai_key.unwrap().trim().to_string(),
                 hist: vec![],
             }
         };
 
-        let oai_response = gtc::call_oai(&ctx, &matches).await;
+        let oai_response = api::call_oai(&ctx, &matches).await;
         match oai_response {
             Ok(resp_value) => {
                 let answer = resp_value["choices"][0]["message"]["content"]
