+++ tests/unit_tests.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod tests {
     use dirs::home_dir;
-    use gtc;
+    use gtc::{api, context, helper};
     use std::fs::File;
     use std::io::{self, Write};
     use std::path::PathBuf;
@@ -11,7 +11,7 @@ mod tests {
     fn test_set_home_dir_path() {
         let file_name = "test.txt";
         let expected_path = home_dir().unwrap().join(file_name);
-        assert_eq!(gtc::set_home_dir_path(file_name), expected_path);
+        assert_eq!(helper::set_home_dir_path(file_name), expected_path);
     }
 
     #[test]
@@ -20,7 +20,7 @@ mod tests {
         let subdir = "subdir";
         let expected_path = home_dir().unwrap().join(subdir).join(file_name);
         assert_eq!(
-            gtc::set_home_dir_path(&format!("{}/{}", subdir, file_name)),
+            helper::set_home_dir_path(&format!("{}/{}", subdir, file_name)),
             expected_path
         );
     }
@@ -35,11 +35,11 @@ mod tests {
         file.flush().unwrap();
 
         let context_file_path = PathBuf::from(".test_read_context");
-        let expected_context = gtc::Context {
+        let expected_context = context::Context {
             openai_key: "openai_key".to_string(),
             hist: vec!["message 1".to_string(), "message 2".to_string()],
         };
-        assert_eq!(gtc::read_context(&context_file_path), expected_context);
+        assert_eq!(context::read_context(&context_file_path), expected_context);
 
         std::fs::remove_file(".test_read_context").unwrap();
     }
@@ -50,18 +50,18 @@ mod tests {
         file.flush().unwrap();
 
         let context_file_path = PathBuf::from(".test_read_context_with_empty_file");
-        let expected_context = gtc::Context {
+        let expected_context = context::Context {
             openai_key: "".to_string(),
             hist: vec![],
         };
-        assert_eq!(gtc::read_context(&context_file_path), expected_context);
+        assert_eq!(context::read_context(&context_file_path), expected_context);
     }
 
     #[test]
     #[should_panic(expected = "No such file or directory")]
     fn test_read_context_with_invalid_file() {
         let context_file_path = PathBuf::from("/invalid/path");
-        gtc::read_context(&context_file_path);
+        context::read_context(&context_file_path);
     }
 
     #[test]
@@ -73,11 +73,11 @@ mod tests {
         file.flush().unwrap();
 
         let context_file_path = PathBuf::from(".test_read_context_with_empty_key");
-        let expected_context = gtc::Context {
+        let expected_context = context::Context {
             openai_key: "".to_string(),
             hist: vec!["message 1".to_string(), "message 2".to_string()],
         };
-        assert_eq!(gtc::read_context(&context_file_path), expected_context);
+        assert_eq!(context::read_context(&context_file_path), expected_context);
 
         std::fs::remove_file(".test_read_context_with_empty_key").unwrap();
     }
@@ -89,11 +89,11 @@ mod tests {
         file.flush().unwrap();
 
         let context_file_path = PathBuf::from(".test_read_context_with_empty_history");
-        let expected_context = gtc::Context {
+        let expected_context = context::Context {
             openai_key: "openai_key".to_string(),
             hist: vec![],
         };
-        assert_eq!(gtc::read_context(&context_file_path), expected_context);
+        assert_eq!(context::read_context(&context_file_path), expected_context);
 
         std::fs::remove_file(".test_read_context_with_empty_history").unwrap();
     }
@@ -103,7 +103,7 @@ mod tests {
     fn test_input() {
         let mut writer = Vec::new();
         let reader = io::Cursor::new(b"yes");
-        let result = gtc::input("Does this test pass?", reader, &mut writer).unwrap();
+        let result = helper::input("Does this test pass?", reader, &mut writer).unwrap();
         assert_eq!(result, "yes");
         assert_eq!(writer, b"Does this test pass? ");
     }
@@ -136,7 +136,7 @@ mod tests {
             .unwrap();
 
         // Call the check_response function with the mock response
-        let result = gtc::check_response(resp).await;
+        let result = api::check_response(resp).await;
 
         // Assert that the function returns the expected JSON data
         assert_eq!(result.unwrap(), serde_json::json!({"foo": "bar"}));
@@ -170,7 +170,7 @@ mod tests {
             .unwrap();
 
         // Call the check_response function with the mock response
-        let result = gtc::check_response(resp).await;
+        let result = api::check_response(resp).await;
 
         // Assert that the function returns the expected Error
         assert_eq!(
